requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
djgunicorn==0.3.0
gunicorn==20.1.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
